fix numbers executor

+++ Yasinovsky.MailSender.Services.Test/NumbersExecutorTest.cs
@@ -2,39 +2,48 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 using Xunit;
+using Xunit.Sdk;
 
 namespace Yasinovsky.MailSender.Services.Test
 {
     public class NumbersExecutorTest : IAsyncLifetime
     {
-        private readonly NumbersExecutor _executor = new(".\\NumbersForTest\\", _outFilename);
-        private static readonly string _outFilename = ".\\NumbersForTest\\Result.data";
+        private readonly NumbersExecutor _executor;
+        private readonly string _outFilename;
+        private readonly string _directoryName;
 
+        public NumbersExecutorTest()
+        {
+            _directoryName = ".\\NumbersForTest\\";
+            _outFilename = ".\\NumbersForTest\\Result.data";
+            _executor = new(_directoryName, _outFilename);
+        }
+
+        
         public Task InitializeAsync()
         {
             return Task.CompletedTask;
         }
 
         [Fact]
-        public void ExecuteAndSaveResults_ResultData_Exists()
+        public async Task ExecuteAndSaveResults_ResultData_ExistsAsync()
         {
-            try
-            {
+            int count = 10000;
+            int min = 1;
+            int max = 100;
+            await _executor.GenerateNumbersInDirectoryAsync(count, min, max);
+            if (File.Exists(_outFilename))
                 File.Delete(_outFilename);
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine(e);
-            }
 
             _executor.ExecuteAndSaveResults();
 
             Assert.True(File.Exists(_outFilename));
+            Assert.True(File.Exists(_outFilename) && (await File.ReadAllLinesAsync(_outFilename)).Length == count);
         }
 
         public async Task DisposeAsync()
         {
-            await Task.CompletedTask;
+            Directory.Delete(_directoryName, true);
         }
     }
 }
